split out decode to seperate function / file
